fix: resolve MSVC C4244/C4100 warnings
- security.cpp: use WideCharToMultiByte for wchar_t to string conversion
- human_module.cpp: add static_cast<int> for int64_t to int conversions
- http_module.cpp: fix size_t to int cast for maxRedirects
- gdi_capture.cpp: suppress unreferenced parameter warning

+++ lib/wingman/src/script/modules/http_module.cpp
@@ -19,7 +19,7 @@ static HttpOptions toHttpOptions(const ScriptValue& v) {
 	auto* followRedirects = v.get("followRedirects");
 	if (followRedirects) options.followRedirects = followRedirects->asBool();
 	auto* maxRedirects = v.get("maxRedirects");
-	if (maxRedirects) options.maxRedirects = static_cast<size_t>(maxRedirects->asInt());
+	if (maxRedirects) options.maxRedirects = static_cast<int>(maxRedirects->asInt());
 	auto* headers = v.get("headers");
 	if (headers && headers->isObject()) {
 		for (const auto& [key, val] : headers->objectVal) {
